Se separaron las clases y lógica para gestionar equipos, técnicos, tareas y ubicaciones en el sistema de mantenimiento, y tener un modelo mucho más ordenado
