+++ modelo/Entidades/TipoMantenimiento.py
@@ -1,0 +1,8 @@
+from enum import Enum, auto
+
+class TipoMantenimiento(Enum):
+    """
+    Enum que define los tipos de mantenimiento disponibles.
+    """
+    PREVENTIVO = auto()
+    CORRECTIVO = auto()
