+++ vista/forms/ubicacion_form.py
@@ -1,4 +1,5 @@
 import tkinter as tk
+from random import randint
 from tkinter import ttk, messagebox
 from typing import Callable
 
@@ -19,10 +20,6 @@ class UbicacionForm:
         frame.pack(fill=tk.BOTH, expand=True)
 
         # Campos del formulario
-        ttk.Label(frame, text="ID Ubicación:").grid(row=0, column=0, sticky=tk.W, pady=5)
-        self.id_entry = ttk.Entry(frame)
-        self.id_entry.grid(row=0, column=1, sticky=tk.EW, pady=5)
-
         ttk.Label(frame, text="Nombre:").grid(row=1, column=0, sticky=tk.W, pady=5)
         self.nombre_entry = ttk.Entry(frame)
         self.nombre_entry.grid(row=1, column=1, sticky=tk.EW, pady=5)
@@ -43,7 +40,7 @@ class UbicacionForm:
     def _guardar(self):
         try:
             ubicacion = self.gestor.registrar_ubicacion(
-                id=self.id_entry.get().strip(),
+                id=self._id_aleatorio(),
                 nombre=self.nombre_entry.get().strip(),
                 descripcion=self.desc_entry.get().strip()
             )
@@ -52,3 +49,6 @@ class UbicacionForm:
             self.window.destroy()
         except ValueError as e:
             messagebox.showerror("Error", f"Datos inválidos: {e}")
+
+    def _id_aleatorio(self):
+        return str(randint(100000, 999999))
